+++ LibCarla/source/carla/client/Vehicle.cpp
@@ -144,5 +144,9 @@ namespace client {
         BaseJSONPath);
   }
 
+  rpc::VehicleFailureState Vehicle::GetFailureState() const {
+    return GetEpisode().Lock()->GetActorSnapshot(*this).state.vehicle_data.failure_state;
+  }
+
 } // namespace client
 } // namespace carla