+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Latest
 
+  * Added failure state to vehicles, which can be retrieved by using `Vehicle.get_failure_state()`. Only Rollover failure state is currently supported.
   * Fixed bug causing the TM to block the simulation when another client teleported a vehicle with no physics.
   * Fixed bug causing the TM to block the simulation when travelling through a short roads that looped on themselves.
   * Improved the TM's handling of non signalized junctions, resulting in a more fluid overall behavior.